refactor: change to atomic structure
