+++ components/atoms/button/BackButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TouchableOpacity, Text, StyleSheet } from 'react-native';
 import {useNavigation} from '@react-navigation/native';
-import { useLanguage } from '../LanguageProvider';
+import { useLanguage } from '../../../utils/LanguageService';
 
 const BackButton = () => {
   const navigation = useNavigation();
