+++ components/organisms/modal/DistanceModal.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Modal, View, Text, StyleSheet, TouchableOpacity } from "react-native";
-import { useLanguage } from "../LanguageProvider";
+import { useLanguage } from "../../../utils/LanguageService";
 
 const DistanceModal = ({
   visible,
@@ -13,7 +13,7 @@ const DistanceModal = ({
   refillMessage,
   distancePerTank,
 }) => {
-  const { language, translations } = useLanguage();
+  const { translations } = useLanguage();
 
   return (
     <Modal visible={visible} animationType="slide">
