+++ navigation/AppNavigator.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
-import CustomHeader from "./CustomHeader";
-import HomePage from "./HomePage";
-import DistancePage from "./DistancePage";
-import FillUpPage from "./FillUpPage";
+import CustomHeader from "../components/organisms/header/Header";
+import HomePage from "../components/pages/HomePage";
+import DistancePage from "../components/pages/DistancePage";
+import FillUpPage from "../components/pages/FillUpPage"
 
 const Stack = createStackNavigator();
