+++ utils/lang/no.js
